+++ CHANGELOG.md
@@ -15,6 +15,17 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+### Changed
+- `com.exonum.binding.storage.indices.MapEntry` moved to package
+  `com.exonum.binding.common.collect`. `FlatMapProof` and `MapIndex` are updated 
+  to use this implementation of `MapEntry`.
+
+### Removed
+- `com.exonum.binding.common.proofs.map.MapEntry` — moved to package
+  `com.exonum.binding.common.collect`.
+
 ## [0.3] - TBA
 
 ### Added